+++ src/main/java/de/fhdw/endpoints/GenreImpl.java
@@ -2,10 +2,13 @@ package de.fhdw.endpoints;
 
 import de.fhdw.models.Artist;
 import de.fhdw.models.Genre;
+import org.jboss.resteasy.annotations.cache.Cache;
 
 import javax.annotation.security.RolesAllowed;
+import javax.persistence.Cacheable;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import javax.ws.rs.core.SecurityContext;
 import java.util.List;
 
@@ -16,55 +19,59 @@ public class GenreImpl implements GenreInterface {
     @Override
     @GET
     @Path("{id}")
-    public Genre get(long id) throws Exception {
-        try {
-            return Genre.findById(id);
-        } catch (Exception e) {
-            throw new Exception("id not found");
+    public Genre get(long id) {
+
+        Genre genre = Genre.findById(id);
+        if (genre == null) {
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
+
+        return Genre.findById(id);
     }
 
     @Override
-    public List<Genre> get() throws Exception {
+    @Cache
+    public List<Genre> get() {
         return Genre.listAll();
     }
 
     @Override
     @PUT
     @RolesAllowed({"employee", "admin"})
-    public Genre put(Genre genre, SecurityContext securityContext) throws Exception {
-        try {
-            Genre old = Artist.findById(genre.id);
-            old.name = genre.name;
-            return old;
-        } catch (Exception E) {
-            throw new Exception("Fehlgeschlagen");
+    public Genre put(Genre genre, SecurityContext securityContext) {
+        Genre old = Artist.findById(genre.id);
+        if (old == null) {
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
+        old.name = genre.name;
+        return old;
     }
 
     @Override
     @POST
     @RolesAllowed({"employee", "admin"})
-    public Genre post(Genre genre, SecurityContext securityContext) throws Exception {
+    public Genre post(Genre genre, SecurityContext securityContext) {
         try {
             genre.persist();
             return genre;
         } catch (Exception e) {
-            throw new Exception("could not persist user");
+            throw new WebApplicationException(Response.Status.NOT_ACCEPTABLE);
         }
     }
 
     @Override
     @DELETE
     @RolesAllowed({"employee", "admin"})
-    public Boolean delete(Genre genre, SecurityContext securityContext) throws Exception {
-        Genre deletedID;
+    public Boolean delete(Genre genre, SecurityContext securityContext)  {
+        Genre deletedID = Genre.findById(genre.id);
+        if(deletedID == null){
+            throw new WebApplicationException(Response.Status.NOT_FOUND);
+        }
         try {
-            deletedID = Genre.findById(genre.id);
             deletedID.delete();
             return true;
         } catch (Exception e) {
-            throw new Exception("User does not exist");
+            throw new WebApplicationException(Response.Status.EXPECTATION_FAILED);
         }
     }
 }
